% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_read.R
\name{fastrbindsf}
\alias{fastrbindsf}
\title{Rapidly bind sf data frames}
\usage{
fastrbindsf(x, check = FALSE)
}
\arguments{
\item{x}{List of sf data frames}

\item{check}{Check that CRS and column names are the same?}
}
\value{
sf data frame
}
\description{
See https://gist.github.com/kadyb/8a32c432a7b497b697379b483a8216d1
And https://github.com/r-spatial/sf/issues/798
}
