% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_read.R
\name{mm_read}
\alias{mm_read}
\alias{mm_list}
\title{Import a directory containing MasterMap data}
\usage{
mm_read(directory = ".", f = NULL, pattern = "Road_FULL", n_files = 1)

mm_list(directory, pattern = "Road_FULL", full.names = FALSE)
}
\arguments{
\item{directory}{The directory of data files to read}

\item{f}{Full file names to read-in}

\item{pattern}{Text string identifying which parts of the road network
to import. Options including FerryLink, FerryNode, FerryTerminal,
Road_FULL, RoadJunctio, RoadLink_FU, RoadNode_FU, and Street_FULL}

\item{n_files}{How many files to read? 1 by default.}

\item{full.names}{Show the full file name? No, meaning just file names
is default.}
}
\value{
sf data frame
}
\description{
See  https://beta.ordnancesurvey.co.uk/products/os-mastermap-highways-network-roads#technical
for technical details (Highway data).
}
\details{
RoadLink, RoadNode, and Street layers have geometries.
}
\examples{
if(FALSE) {
# Change directory to the location where you store data:
directory = "path/to/data"
files = mm_list(directory, pattern = "*")
# file_strings = stringr::str_sub(files, 16, 26)
# [1] ""            "FerryLink_F" "FerryNode_F" "FerryTermin"
# [5] "Road_FULL_0" "RoadJunctio" "RoadLink_FU" "RoadNode_FU"
# [9] "Street_FULL"
# unique(file_strings)
files = mm_list(directory, pattern = "RoadLink")
head(files)
mm_data_small = mm_read(directory, pattern = "RoadLink")
nrow(mm_data_small) # 46000
sf::st_write(mm_data_small, "mm_data_small.gpkg")
names(mm_data)
plot(mm_data$centrelineGeometry)
# Multiple files
mm_data = mm_read(directory, pattern = "RoadLink", n_files = 3)
mm_data
plot(mm_data$centrelineGeometry)
}
}
