% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mastermap.R
\name{rbind_list_fill}
\alias{rbind_list_fill}
\title{Return all list elements with consistent column names}
\usage{
rbind_list_fill(x)
}
\arguments{
\item{x}{A list of data frames}
}
\description{
Return all list elements with consistent column names
}
