% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mastermap.R
\name{read_mastermap}
\alias{read_mastermap}
\title{Read mastermap data}
\usage{
read_mastermap(dir, type_match = "roadlink", n = 5)
}
\arguments{
\item{dir}{Directory where MasterMap files live}

\item{type_match}{Type of feature to return, matched on directory}

\item{n}{How many files to return (all by default, can be resource-consuming)}
}
\description{
Read mastermap data
}
