% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mastermap.R
\name{read_mastermap}
\alias{read_mastermap}
\title{Read mastermap data}
\usage{
read_mastermap(dir, type_match = "roadlink", n = 10000, ext = "gz")
}
\arguments{
\item{dir}{Directory where MasterMap files live}

\item{type_match}{Type of feature to return, matched on directory}

\item{n}{How many files to return (all by default, can be resource-consuming)}

\item{ext}{Extension, e.g. .gz}
}
\description{
This function can read-in multiple files stored in datasets downloaded from the Ordnance Survey.
It was originally designed to read-in
\href{https://www.ordnancesurvey.co.uk/documents/os-mastermap-highways-network-product-guide.pdf}{MasterMap Highway data}.
}
\examples{
\donttest{
dir = "~/hd/data/os/Download_mastermap-roads-2019_1483661/MasterMap Highways Network_rami_3480319/"
roads_uk = read_mastermap(dir, n = 2)
roads_uk = read_mastermap(dir)
}
}
